+++ contracts/token-native-ft/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod contract;
 pub mod error;
+pub mod msg;
 pub mod proto;
 pub mod serde;
 pub mod state;
